pyproject.toml
[project]
name = "viora-ai-personal-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9,<4.0"
dependencies = [
    "elevenlabs>=2.8.1",
    "gradio>=4.44.1",
    "groq>=0.30.0",
    "gtts>=2.5.4",
    "langchain-google-genai>=2.1.8",
    "langgraph>=0.5.4",
    "opencv-python>=4.12.0.88",
    "playsound==1.2.2",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.1",
    "speechrecognition>=3.14.3",
]
